COMMERCE-1277 Update component UI - Server URL is pulled up in the configuration hierarchy
