+++ modules/etl/talend/src/main/resources/com/liferay/talend/configuration/Messages.properties
@@ -1,3 +1,3 @@
-LiferayInputMapperConfiguration._maxItemsPerRequest._displayName=Max Items Per Request
+LiferayInputMapperConfiguration._batchSize._displayName=Batch Size
 LiferayInputMapperConfiguration._timeout._displayName=Timeout
 LiferayInputMapperConfiguration._restDataSet._displayName=REST Data Set
