+++ modules/etl/talend/src/main/resources/com/liferay/talend/datastore/Messages.properties
@@ -2,11 +2,10 @@ AuthenticationMethod.BASIC._displayName=Basic
 AuthenticationMethod.OAUTH2._displayName=OAuth2
 BasicDataStore._anonymous._displayName=Anonymous
 BasicDataStore._password._displayName=Password
-BasicDataStore._serverURL._displayName=ServerURL
+InputDataStore._serverURL._displayName=ServerURL
 BasicDataStore._user._displayName=User
 InputDataStore._authenticationMethod._displayName =Authentication Method
 InputDataStore._basicDataStore._displayName =Basic Data Store
 InputDataStore._oAuthDataStore._displayName =OAuth2 Data Store
 OAuthDataStore._consumerKey._displayName =Consumer Key
 OAuthDataStore._consumerSecret._displayName =Consumer Secret
-OAuthDataStore._serverURL._displayName =Server URL
